update interactors to new names
