+++ shared/src/business/useCases/caseNote/deleteCaseNoteInteractor.js
@@ -2,6 +2,7 @@ const {
   isAuthorized,
   ROLE_PERMISSIONS,
 } = require('../../../authorization/authorizationClientService');
+const { Case } = require('../../entities/cases/Case');
 const { UnauthorizedError } = require('../../../errors/errors');
 
 /**
@@ -9,23 +10,32 @@ const { UnauthorizedError } = require('../../../errors/errors');
  *
  * @param {object} providers the providers object
  * @param {object} providers.applicationContext the application context
- * @param {string} providers.caseId the id of the case the case note is attached to
+ * @param {string} providers.caseId the id of the case the procedural note is attached to
  * @returns {Promise} the promise of the delete call
  */
 exports.deleteCaseNoteInteractor = async ({ applicationContext, caseId }) => {
   const user = applicationContext.getCurrentUser();
 
-  if (!isAuthorized(user, ROLE_PERMISSIONS.TRIAL_SESSION_WORKING_COPY)) {
+  if (!isAuthorized(user, ROLE_PERMISSIONS.PROCEDURAL_NOTES)) {
     throw new UnauthorizedError('Unauthorized');
   }
+  const caseRecord = await applicationContext
+    .getPersistenceGateway()
+    .getCaseByCaseId({
+      applicationContext,
+      caseId,
+    });
 
-  const judgeUser = await applicationContext
-    .getUseCases()
-    .getJudgeForUserChambersInteractor({ applicationContext, user });
-
-  return await applicationContext.getPersistenceGateway().deleteCaseNote({
+  delete caseRecord.proceduralNote;
+  const caseToUpdate = new Case(caseRecord, {
     applicationContext,
-    caseId,
-    userId: judgeUser.userId,
+  })
+    .validate()
+    .toRawObject();
+
+  const result = await applicationContext.getPersistenceGateway().updateCase({
+    applicationContext,
+    caseToUpdate,
   });
+  return result;
 };
