+++ shared/src/business/useCases/caseNote/deleteJudgesCaseNoteInteractor.js
@@ -1,0 +1,34 @@
+const {
+  isAuthorized,
+  ROLE_PERMISSIONS,
+} = require('../../../authorization/authorizationClientService');
+const { UnauthorizedError } = require('../../../errors/errors');
+
+/**
+ * deleteJudgesCaseNoteInteractor
+ *
+ * @param {object} providers the providers object
+ * @param {object} providers.applicationContext the application context
+ * @param {string} providers.caseId the id of the case the case note is attached to
+ * @returns {Promise} the promise of the delete call
+ */
+exports.deleteJudgesCaseNoteInteractor = async ({
+  applicationContext,
+  caseId,
+}) => {
+  const user = applicationContext.getCurrentUser();
+
+  if (!isAuthorized(user, ROLE_PERMISSIONS.TRIAL_SESSION_WORKING_COPY)) {
+    throw new UnauthorizedError('Unauthorized');
+  }
+
+  const judgeUser = await applicationContext
+    .getUseCases()
+    .getJudgeForUserChambersInteractor({ applicationContext, user });
+
+  return await applicationContext.getPersistenceGateway().deleteJudgesCaseNote({
+    applicationContext,
+    caseId,
+    userId: judgeUser.userId,
+  });
+};
