+++ shared/src/business/useCases/caseNote/getJudgesCaseNoteInteractor.js
@@ -6,18 +6,16 @@ const { CaseNote } = require('../../entities/cases/CaseNote');
 const { UnauthorizedError } = require('../../../errors/errors');
 
 /**
- * updateCaseNoteInteractor
+ * getJudgesCaseNoteInteractor
  *
  * @param {object} providers the providers object
  * @param {object} providers.applicationContext the application context
- * @param {string} providers.caseId the id of the case to update notes
- * @param {string} providers.notes the notes to update
- * @returns {object} the updated case note returned from persistence
+ * @param {string} providers.caseId the id of the case to get notes for
+ * @returns {object} the case note object if one is found
  */
-exports.updateCaseNoteInteractor = async ({
+exports.getJudgesCaseNoteInteractor = async ({
   applicationContext,
   caseId,
-  notes,
 }) => {
   const user = applicationContext.getCurrentUser();
   if (!isAuthorized(user, ROLE_PERMISSIONS.TRIAL_SESSION_WORKING_COPY)) {
@@ -28,18 +26,15 @@ exports.updateCaseNoteInteractor = async ({
     .getUseCases()
     .getJudgeForUserChambersInteractor({ applicationContext, user });
 
-  const caseNoteEntity = new CaseNote({
-    caseId,
-    notes,
-    userId: judgeUser.userId,
-  });
+  const caseNote = await applicationContext
+    .getPersistenceGateway()
+    .getJudgesCaseNote({
+      applicationContext,
+      caseId,
+      userId: judgeUser.userId,
+    });
 
-  const caseNoteToUpdate = caseNoteEntity.validate().toRawObject();
-
-  await applicationContext.getPersistenceGateway().updateCaseNote({
-    applicationContext,
-    caseNoteToUpdate,
-  });
-
-  return caseNoteToUpdate;
+  if (caseNote) {
+    return new CaseNote(caseNote).validate().toRawObject();
+  }
 };
